大量优化结构

+++ src/main/java/com/ucstu/guangbt/djzhaopin/model/PageResult.java
@@ -18,7 +18,7 @@ import lombok.experimental.Accessors;
 @Accessors(chain = true)
 public class PageResult<T> {
 
-    private Integer totalCount;
+    private long totalCount;
 
     @JsonIgnore
     private List<T> contents;
